# swagger.yaml
openapi: "3.0.0"
info:
  version: "1.0.0"
  title: "Movies API"
  description: "This is a simple API for managing a collection of movies."
paths:
  /movies:
    get:
      summary: "Get all movies"
      responses:
        '200':
          description: "A list of movies"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Movie"
    post:
      summary: "Create a new movie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movie"
      responses:
        '201':
          description: "The created movie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
  /movies/{id}:
    get:
      summary: "Get a movie by ID"
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: "Movie ID"
      responses:
        '200':
          description: "A movie object"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        '404':
          description: "Movie not found"
    put:
      summary: "Update a movie by ID"
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: "Movie ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movie"
      responses:
        '200':
          description: "The updated movie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        '404':
          description: "Movie not found"
    delete:
      summary: "Delete a movie by ID"
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: "Movie ID"
      responses:
        '200':
          description: "The remaining movies"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Movie"
        '404':
          description: "Movie not found"
components:
  schemas:
    Movie:
      type: "object"
      properties:
        id:
          type: "string"
        genre:
          type: "string"
        title:
          type: "string"
        director:
          $ref: "#/components/schemas/Director"
    Director:
      type: "object"
      properties:
        first_name:
          type: "string"
        last_name:
          type: "string"